move styles folder
