+++ pages/_app.tsx
@@ -1,10 +1,10 @@
-import '../src/styles/css/fontawesome-all.min.css'
-import '../src/styles/css/main.css'
-import '../src/styles/css/noscript.css'
-import '../src/styles/scss/variables/_breakpoints.scss'
-import '../src/styles/scss/variables/_colors.scss'
-import '../src/styles/scss/_normalizer.scss'
-import '../src/styles/scss/application.scss'
+import '../styles/css/fontawesome-all.min.css'
+import '../styles/css/main.css'
+import '../styles/css/noscript.css'
+import '../styles/scss/variables/_breakpoints.scss'
+import '../styles/scss/variables/_colors.scss'
+import '../styles/scss/_normalizer.scss'
+import '../styles/scss/application.scss'
 
 import type { AppProps } from 'next/app'
 import { Provider } from 'react-redux'
